Refactor codebase
Split into lib, CLI, and GUI for better organisation.
