+++ Cargo.toml
@@ -1,46 +1,50 @@
 [package]
 build = "src/build.rs"
 name = "yanu"
-version = "0.6.1"
+version.workspace = true
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
-[features]
-# ! Intended only for a aarch64 linux env in termux
-android-proot = []
+[workspace]
+members = ["libyanu-common", "yanu-cli"]
 
-[dependencies]
-eyre = "0.6"
+[workspace.package]
+version = "0.6.1"
+
+[workspace.dependencies]
 color-eyre = "0.6.2"
-dirs = "5.0"
-inquire = "0.6"
-serde = { version = "1.0", features = ["derive"] }
-confy = { version = "0.5", default-features = false, features = ["ron_conf"] }
-clap = { version = "4.1", features = ["derive"] }
-tracing-subscriber = "0.3"
+eyre = "0.6"
 tracing = "0.1"
-hex = "0.4"
-tempfile = "3.4.0"
+tracing-subscriber = "0.3"
 tracing-appender = "0.2"
-strum = "0.24"
-strum_macros = "0.24"
-walkdir = "2.3"
-ctrlc = { version = "3.2", features = ["termination"] }
-indicatif = "0.17"
-console = "0.15"
-once_cell = "1.17.1"
+tempfile = "3.4"
 fs-err = "2.9.0"
-
-[target.'cfg(unix)'.dependencies]
+serde = { version = "1.0", features = ["derive"] }
+walkdir = "2.3"
 tilde-expand = "0.1.1"
 
-[target.'cfg(any(target_os = "linux", target_os = "windows"))'.dependencies]
-# Switch to main branch once zenity is merged
+[dependencies]
+eyre = { workspace = true }
+color-eyre = { workspace = true }
+dirs = "5.0"
+tracing = { workspace = true }
+tracing-subscriber = { workspace = true }
+tracing-appender = { workspace = true }
+tempfile = { workspace = true }
+fs-err = { workspace = true }
+walkdir = { workspace = true }
+#* Switch to main branch once zenity is merged
 rfd = { git = "https://github.com/PolyMeilex/rfd.git", rev = "52d1010b09a957d3f1e40931f154805e5e1e57a6", default_features = false, features = [
     "xdg-portal",
 ] }
-winapi = "0.3.9"
+libyanu-common = { path = "libyanu-common" }
+
+[target.'cfg(unix)'.dependencies]
+tilde-expand = { workspace = true }
+
+[target.'cfg(windows)'.dependencies]
+winapi = { version = "0.3.9", features = ["winuser"] }
 
 [profile.release]
 strip = true
