+++ libyanu-common/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "libyanu-common"
+version.workspace = true
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[features]
+# ! Intended only for a aarch64-linux target in termux
+android-proot = []
+
+[dependencies]
+eyre = { workspace = true }
+color-eyre = { workspace = true }
+confy = { version = "0.5", default-features = false, features = ["ron_conf"] }
+dirs = "5.0"
+serde = { workspace = true }
+tracing = { workspace = true }
+tracing-subscriber = { workspace = true }
+hex = "0.4"
+strum = "0.24"
+strum_macros = "0.24"
+once_cell = "1.17"
+walkdir = { workspace = true }
+fs-err = { workspace = true }
+tempfile = { workspace = true }
