+++ libyanu-common/src/hac/ticket.rs
