+++ libyanu-common/src/termux.rs
