+++ src/build.rs
@@ -1,4 +1,4 @@
 fn main() {
     #[cfg(target_os = "windows")]
-    embed_resource::compile("resources/manifest.rc");
+    embed_resource::compile("resources/manifest.rc", embed_resource::NONE);
 }
