--- src/cli/mod.rs
@@ -1,1 +1,0 @@
-pub mod args;
